Extract probability utilities into calc module
