+++ calc/__init__.py
@@ -1,0 +1,8 @@
+from .core import indiv_prob, group_prob, fights_needed, cumulative
+
+__all__ = [
+    "indiv_prob",
+    "group_prob",
+    "fights_needed",
+    "cumulative",
+]
